* master:
  fix: bump minimum alby migration balance to 200,000 to account for channel fees
  deploy templates (#276)
  feat: dynamic wallet sync intervals based on user activity (#330)
  fix: add legal copy (#277)
  feat: connect some peers at LDK node startup for p2p gossip (#326)
  feat: update ldk-node version, add button in debug tools to fetch network graph (#327)
  feat: manually link account (#319)
  fix: settings layout on mobile
  feat: set global properties on event publisher
  feat: reset router when backing up node, update ldk-node dependency (#317)
  fix: LDK shutdown (#313)
  Start NWC after the network is actually online
  fix: remove replace LDK node stop with destroy
  fix: LDK add node.Destroy after node.Stop
  fix: update copy of import button
  fix: use Stop instead of Destroy in LDK node
  chore: add some more logs when shutting down the node
  fix: copy
  Dependency updates 20240514 (#310)
  feat: logout / shutdown (#292)

+++ frontend/src/screens/channels/Channels.tsx
@@ -48,6 +48,7 @@ import { useChannels } from "src/hooks/useChannels";
 import { useInfo } from "src/hooks/useInfo";
 import { useNodeConnectionInfo } from "src/hooks/useNodeConnectionInfo.ts";
 import { useRedeemOnchainFunds } from "src/hooks/useRedeemOnchainFunds.ts";
+import { useSyncWallet } from "src/hooks/useSyncWallet.ts";
 import { copyToClipboard } from "src/lib/clipboard.ts";
 import { formatAmount } from "src/lib/utils.ts";
 import { CloseChannelResponse, Node } from "src/types";
@@ -55,6 +56,7 @@ import { request } from "src/utils/request";
 import { useCSRF } from "../../hooks/useCSRF.ts";
 
 export default function Channels() {
+  useSyncWallet();
   const { data: channels, mutate: reloadChannels } = useChannels();
   const { data: nodeConnectionInfo } = useNodeConnectionInfo();
   const { data: balances } = useBalances();
@@ -126,10 +128,21 @@ export default function Channels() {
         return;
       }
 
+      const closeType = prompt(
+        "Select way to close the channel. Type 'force close' if you want to force close the channel. Note: your channel balance will be locked for up to two weeks if you force close.",
+        "normal close"
+      );
+      if (!closeType) {
+        console.error("Cancelled close channel");
+        return;
+      }
+
       console.log(`🎬 Closing channel with ${nodeId}`);
 
       const closeChannelResponse = await request<CloseChannelResponse>(
-        `/api/peers/${nodeId}/channels/${channelId}`,
+        `/api/peers/${nodeId}/channels/${channelId}?force=${
+          closeType === "force close"
+        }`,
         {
           method: "DELETE",
           headers: {
@@ -143,12 +156,17 @@ export default function Channels() {
         throw new Error("Error closing channel");
       }
 
-      await reloadChannels();
-
-      console.log(
-        "Closed channel",
-        channels?.find((c) => c.id === channelId && c.remotePubkey === nodeId)
+      const closedChannel = channels?.find(
+        (c) => c.id === channelId && c.remotePubkey === nodeId
       );
+      console.log("Closed channel", closedChannel);
+      if (closedChannel) {
+        prompt(
+          "Closed channel. Copy channel funding TX to view on mempool",
+          closedChannel.fundingTxId
+        );
+      }
+      await reloadChannels();
       toast({ title: "Sucessfully closed channel." });
     } catch (error) {
       console.error(error);
